+++ src/main/java/com/web/stard/domain/member/dto/request/MemberRequestDto.java
@@ -118,4 +118,14 @@ public class MemberRequestDto {
         private String nickname;
     }
 
+    @Getter
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class NicknameDto {
+        @Size(min = 2, message = "닉네임은 2자 이상이어야 합니다.")
+        @Schema(example = "스타", description = "회원 닉네임")
+        @NotBlank(message = "닉네임은 필수 입니다.")
+        private String nickname;
+    }
+
 }
